requirement.txt
openai==0.27.2
fastapi==0.95.0
uvicorn==0.21.1
pandas==1.5.3
matplotlib==3.7.1
plotly==5.13.1
scipy==1.10.1
scikit-learn==1.2.2
pypdf==3.6.0
python-dotenv==1.0.0
python-multipart==0.0.6